Add files via upload
In this lab, we aim to compare Shell and Heap Sorting methods to obtain some appreciation for the importance of selecting the right sorting method for a given data set. In this project, I went for an iterative approach for both methods.  As a structure, array, as well as Linked array, were used to store values of the data sets as well as the run time to calculate the differences between the sorting methods. Linked arrays provide a better use of space and are easy to access and store. Arrays provide a structure where values of each ordered set can be stored, retrieved, and ordered in a convenient manner. In this example, we are not worried about the non-flexibility of the array, because the sizes of the files are preset and will not be altered at any point.